Update location text fallback and test it (#7408)
