+++ src/i18n/strings/en_EN.json
@@ -1666,7 +1666,6 @@
     "%(userName)s (power %(powerLevelNumber)s)": "%(userName)s (power %(powerLevelNumber)s)",
     "Send message": "Send message",
     "Add emoji": "Add emoji",
-    "Share location": "Share location",
     "Upload file": "Upload file",
     "You do not have permission to start polls in this room.": "You do not have permission to start polls in this room.",
     "Create poll": "Create poll",
@@ -2121,6 +2120,7 @@
     "edited": "edited",
     "Submit logs": "Submit logs",
     "Can't load this message": "Can't load this message",
+    "Share location": "Share location",
     "Share custom location": "Share custom location",
     "Share my current location as a once off": "Share my current location as a once off",
     "My location": "My location",
